template.go containing generic STL-like algorithms.
This is part of an attempt to find the best way to reuse STL-like code in Go
absent the generics/templates that make the Standard Template Library both
possible and efficient in C++. The idea being, copy-and-paste what you need
for your type, performing the appropriate textual substitutions, and tweak as
necessary should performance concerns require. Not necessarily ideal from a
code-reuse point of view, but given these are standard, straightforward
algorithms with permanently-defined behavior and performance characteristics,
it does work for this specific case without further complication to the Go
language itself.