Properly normalize var tokens and top-level tokens
The insight to differentiating top-level tokens starting with './' from tokens
corresponding to files within the Makefile directory is to detect whether or
not the token corresponds to an existing top-level object, or to one of the
'lib*.{a,so,dylib}' objects that will be built there. This means that the
script must always be run from the top-level directory, but I think that's a
given.